# src/payments/payments.controller.ts
import {
  Controller,
  Post,
  Get,
  Param,
  Query,
  UseGuards,
  Req,
  Res,
  HttpCode,
  NotFoundException,
} from '@nestjs/common';
import {
  ApiTags,
  ApiOperation,
  ApiResponse,
  ApiBearerAuth,
} from '@nestjs/swagger';
import { ConfigService } from '@nestjs/config';
import { SkipThrottle } from '@nestjs/throttler';
import type { Request, Response } from 'express';
import { PaymentsService } from './payments.service';
import { PaymentStatus } from './entities/payment.entity';
import { JwtAuthGuard } from '../auth/jwt-auth.guard';
import { CurrentUser } from '../common/decorators/current-user.decorator';
import { UserRole } from '../users/entities/user.entity';

@ApiTags('payments')
@Controller('payments')
export class PaymentsController {
  constructor(
    private readonly paymentsService: PaymentsService,
    private readonly configService: ConfigService,
  ) {}

  @Post(':orderId/initiate')
  @UseGuards(JwtAuthGuard)
  @ApiBearerAuth()
  @ApiOperation({ summary: 'BOG გადახდის დაწყება კონკრეტული შეკვეთისთვის' })
  @ApiResponse({
    status: 201,
    description: 'redirectUrl — მომხმარებელი აქ უნდა გადამისამართდეს',
  })
  @ApiResponse({
    status: 400,
    description: 'შეკვეთა უკვე გადახდილია ან PENDING არაა',
  })
  @ApiResponse({ status: 403, description: 'სხვისი შეკვეთაა' })
  initiate(
    @CurrentUser() user: { userId: number; role: UserRole },
    @Param('orderId') orderId: string,
  ) {
    return this.paymentsService.initiate(user.userId, user.role, +orderId);
  }

  // BOG არაა ავტორიზებული მომხმარებელი — ამ route-ზე JwtAuthGuard არ დგას,
  // დაცვა მხოლოდ ხელმოწერის ვერიფიკაციითაა (PaymentsService.handleCallback
  // 401/403-ს აგდებს არასწორ/არარსებულ Callback-Signature-ზე).
  @Post('callback/bog')
  // ⚠️ გლობალური per-IP rate limit (AppModule) ამ route-ს არ უნდა ეხებოდეს:
  // BOG-ის callback-ები ვიწრო IP-დიაპაზონიდან მოდის, ანუ ერთ „მომხმარებლად"
  // ჩაითვლება — ბურსტზე 429 დაბრუნება იმას ნიშნავს, რომ რეალურად გადახდილი
  // შეკვეთა ჩვენს ბაზაში გადაუხდელი დარჩება. დაცვა ხელმოწერის ვერიფიკაციაა.
  @SkipThrottle()
  @HttpCode(200)
  @ApiOperation({
    summary: 'BOG-ის გადახდის callback (არაავტორიზებული, ხელმოწერით დაცული)',
  })
  async bogCallback(@Req() req: Request) {
    // main.ts-ში rawBody: true ჩართულია — ვერიფიკაცია ზუსტად იმ ბაიტებზე
    // ხდება, რაც BOG-მა მოაწერა ხელი (არა ხელახლა-serialized JSON-ზე).
    const rawBody = (req as Request & { rawBody?: Buffer }).rawBody;
    await this.paymentsService.handleCallback(
      rawBody ?? Buffer.alloc(0),
      req.headers as Record<string, string>,
    );
    return { received: true };
  }

  // MockPaymentProvider-ის "checkout გვერდი" — რეალურ BOG checkout-ს ცვლის
  // კომპანიის რეგისტრაციამდე. მხოლოდ PAYMENT_PROVIDER=mock-ზეა ხელმისაწვდომი,
  // რომ production-ში (PAYMENT_PROVIDER=bog) ვინმემ უფასოდ ვერ "გადაიხადოს".
  @Get('mock/:externalId/complete')
  @ApiOperation({
    summary:
      'MockPaymentProvider-ის auto-complete (მხოლოდ PAYMENT_PROVIDER=mock)',
  })
  async completeMockPayment(
    @Param('externalId') externalId: string,
    @Query('orderId') orderId: string,
    @Res() res: Response,
  ) {
    if (this.configService.get<string>('PAYMENT_PROVIDER') === 'bog') {
      throw new NotFoundException();
    }

    const rawBody = Buffer.from(
      JSON.stringify({ externalId, status: PaymentStatus.COMPLETED }),
      'utf8',
    );
    await this.paymentsService.handleCallback(rawBody, {});

    const frontendUrl = this.configService.get<string>('FRONTEND_URL');
    return res.redirect(`${frontendUrl}/orders/${orderId}?payment=success`);
  }
}