# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/otp/send": {
      "post": {
        "operationId": "OtpController_sendOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "კოდი გაიგზავნა",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendOtpResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "გაგზავნა ვერ მოხერხდა"
          }
        },
        "summary": "OTP კოდის გაგზავნა მითითებულ მობილურზე",
        "tags": [
          "otp"
        ]
      }
    },
    "/otp/verify": {
      "post": {
        "operationId": "OtpController_verifyOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "შედეგი: verified true/false"
          }
        },
        "summary": "OTP კოდის დადასტურება",
        "tags": [
          "otp"
        ]
      }
    },
    "/otp/send-email": {
      "post": {
        "operationId": "OtpController_sendEmailOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendEmailOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "კოდი გაიგზავნა",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendOtpResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "გაგზავნა ვერ მოხერხდა"
          }
        },
        "summary": "OTP კოდის გაგზავნა მითითებულ ელფოსტაზე",
        "tags": [
          "otp"
        ]
      }
    },
    "/otp/verify-email": {
      "post": {
        "operationId": "OtpController_verifyEmailOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "შედეგი: verified true/false"
          }
        },
        "summary": "ელფოსტაზე გაგზავნილი OTP კოდის დადასტურება",
        "tags": [
          "otp"
        ]
      }
    },
...
    "/payments/{orderId}/initiate": {
      "post": {
        "operationId": "PaymentsController_initiate",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "redirectUrl — მომხმარებელი აქ უნდა გადამისამართდეს"
          },
          "400": {
            "description": "შეკვეთა უკვე გადახდილია ან PENDING არაა"
          },
          "403": {
            "description": "სხვისი შეკვეთაა"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "BOG გადახდის დაწყება კონკრეტული შეკვეთისთვის",
        "tags": [
          "payments"
        ]
      }
    },
...
    "/payments/mock/{externalId}/complete": {
      "get": {
        "operationId": "PaymentsController_completeMockPayment",
        "parameters": [
          {
            "name": "externalId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "MockPaymentProvider-ის auto-complete (მხოლოდ PAYMENT_PROVIDER=mock)",
        "tags": [
          "payments"
        ]
      }
    },
...
  },
  "info": {
    "title": "Online Shop API",
    "description": "REST API for the online shop backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SendOtpDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "description": "მობილურის ნომერი, სადაც გაიგზავნება კოდი"
          }
        },
        "required": [
          "phoneNumber"
        ]
      },
      "SendOtpResponseDto": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "description": "verify.ge-ს მიერ დაბრუნებული მოთხოვნის ID"
          }
        },
        "required": [
          "requestId"
        ]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "description": "send-otp-ის პასუხიდან მიღებული requestId"
          },
          "code": {
            "type": "string",
            "description": "მომხმარებლის მობილურზე მიღებული კოდი"
          }
        },
        "required": [
          "requestId",
          "code"
        ]
      },
      "SendEmailOtpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "ელფოსტა, სადაც გაიგზავნება დადასტურების კოდი"
          }
        },
        "required": [
          "email"
        ]
      },
      "VerifyEmailOtpDto": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "description": "send-email-ის პასუხიდან მიღებული requestId"
          },
          "code": {
            "type": "string",
            "description": "მომხმარებლის ელფოსტაზე მიღებული კოდი"
          }
        },
        "required": [
          "requestId",
          "code"
        ]
      },
...
    }
  }
}